Move the history command into it's own module
Plus some imports cleanup.
